Input, Button, Textarea and DialogPopup components have been moved to main components folder
