+++ src/components/AddCustomer.vue
@@ -61,9 +61,9 @@
 
 <script setup>
 import { ref } from 'vue';
-import Input from '@/components/layout/Input.vue';
-import Button from '@/components/layout/Button.vue';
-import Textarea from '@/components/layout/Textarea.vue';
+import Input from '@/components/Input.vue';
+import Button from '@/components/Button.vue';
+import Textarea from '@/components/Textarea.vue';
 
 function consoleClick() {
   console.log('Button clicked');
