+++ src/components/ForgotPassword.vue
@@ -65,6 +65,6 @@
 
 <script setup>
 import logo from '@/assets/logo-browtricks.png';
-import Button from '@/components/layout/Button.vue';
-import Input from '@/components/layout/Input.vue';
+import Button from '@/components/Button.vue';
+import Input from '@/components/Input.vue';
 </script>
